requirements.txt
pytest
pytest-asyncio
anthropic
openai
langchain
langsmith
GitPython
python-dotenv
